Implemented cache system.
Tmdb's api limits to 40 requests every 10 seconds currently. It is a good practice to cache the application responses, to reduce resource consumption and avoid further problems. For example, let's imagine that we implement this API on a website that makes a request to tmdbapi every time a user accesses a page. If a malicious user repeats the same operation 40 times in less than 10 seconds it will cause the website to fall for a few seconds. A simple method would be to store the answers in our database. And update this data periodically. But much simpler would be if we stored the responses of the Api in cache, so as not to overload our database with unnecessary data.

 - Using library functools (python 3+) https://docs.python.org/3/library/functools.html

 - Only method cached is 'cached_request(url)', which calls the function 'request.get(url)'. The cache discriminates according to the 'url' parameter, if is repeated, it will deliver a cached version of the response.

 - If 'cached' param is True in '_call()' method run 'cached_request(url)', else runs 'request.get(url)' without caching. At the moment all requests can be cached, but I did it in case you want to add account functions from TmdbApi in the future. In any case by default the cache parameter will be True

 - Constant REQUEST_CACHE_MAXSIZE will define the maxsize of the cache, None by default (no limit). The problem is that I cannot use a variable to define the maxsize parameter of the lru_cache decorator, then users should adjust this manually. I can't find another solution

 - Another problem to solve is that users should clean the cache periodically if they do not restart their applications. That is why I have added the 'cache_clear()' method to TMDb class, which can be called periodically.

 - I can see in /requeriments.txt that the backport of functools for Python 3.0+ and 2.6+ is already added, but never used. I use Python 3.3+, so I have not tried the backport.

 '''
 backports.functools-lru-cache==1.5
 '''